Refactor listening into pub/sub and add CoHost/ShowAssistant
This effort focuses on allowing subscribing to the Listener. In doing
so, it moves the question listening toggle out of the Listener class.

The question toggle now lives in a new CoHost class that handles
listening for a question and answer it.

Show note-taking, like topic and summarization, now happens in a new
ShowAssistant class. In the future, we will introduce an interface to
allow adding actions that can be taken at specified intervals (like is
currently done with summarization and topic extraction).